requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Authentication
PyJWT==2.8.0
flask-jwt-extended==4.5.3

# Database (uncomment based on your needs)
# SQLAlchemy==2.0.23
# pymongo==4.6.1
# psycopg2-binary==2.9.9  # for PostgreSQL

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
datetime